refactor: remove legacy content and documentation components
