--- content/authors/shadcn.mdx
@@ -1,5 +1,0 @@
----
-title: shadcn
-avatar: /images/avatars/shadcn.png
-twitter: shadcn
----
