--- content/docs/documentation/index.mdx
@@ -1,60 +1,0 @@
----
-title: Documentation
-description: Build your documentation site using Contentlayer and MDX.
----
-
-Taxonomy includes a documentation site built using [Contentlayer](https://contentlayer.dev) and [MDX](http://mdxjs.com).
-
-## Features
-
-It comes with the following features out of the box:
-
-1. Write content using MDX.
-2. Transform and validate content using Contentlayer.
-3. MDX components such as `<Callout />` and `<Card />`.
-4. Support for table of contents.
-5. Custom navigation with prev and next pager.
-6. Beautiful code blocks using `rehype-pretty-code`.
-7. Syntax highlighting using `shiki`.
-8. Built-in search (_in progress_).
-9. Dark mode (_in progress_).
-
-## How is it built
-
-Click on a section below to learn how the documentation site built.
-
-<div className="grid gap-4 mt-6">
-
-<Card href="/docs/documentation">
-
-### Contentlayer
-
-Learn how to use MDX with Contentlayer.
-
-</Card>
-
-<Card href="/docs/documentation/components">
-
-### Components
-
-Using React components in Mardown.
-
-</Card>
-
-<Card href="/docs/documentation/components">
-
-### Code Blocks
-
-Beautiful code blocks with syntax highlighting.
-
-</Card>
-
-<Card href="/docs/documentation/components">
-
-### Style Guide
-
-View a sample page with all the styles.
-
-</Card>
-
-</div>
