--- content/docs/in-progress.mdx
@@ -1,10 +1,0 @@
----
-title: Not Implemented
-description: This page is in progress.
----
-
-<Callout>
-
-This site is a work in progress. If you see dummy text on a page, it means I'm still working on it. You can follow updates on Twitter [@shadcn](https://twitter.com/shadcn).
-
-</Callout>
