--- content/docs/index.mdx
@@ -1,54 +1,0 @@
----
-title: Documentation
-description: Welcome to the Taxonomy documentation.
----
-
-This is the documentation for the Taxonomy site.
-
-This is an example of a doc site built using [ContentLayer](/docs/documentation/contentlayer) and MDX.
-
-<Callout>
-
-This site is a work in progress. If you see dummy text on a page, it means I'm still working on it. You can follow updates on Twitter [@shadcn](https://twitter.com/shadcn).
-
-</Callout>
-
-## Features
-
-Select a feature below to learn more about it.
-
-<div className="grid sm:grid-cols-2 gap-4 mt-6">
-
-<Card href="/docs/documentation">
-
-### Documentation
-
-This documentation site built using Contentlayer.
-
-</Card>
-
-<Card href="/docs/marketing" disabled>
-
-### Marketing
-
-The marketing site with landing pages.
-
-</Card>
-
-<Card href="/docs/app" disabled>
-
-### App
-
-The dashboard with auth and subscriptions.
-
-</Card>
-
-<Card href="/docs/blog" disabled>
-
-### Blog
-
-The blog built using Contentlayer and MDX.
-
-</Card>
-
-</div>
