+++ lib/auth.config.ts
