--- lib/validations/post.ts
@@ -1,8 +1,0 @@
-import * as z from "zod"
-
-export const postPatchSchema = z.object({
-  title: z.string().min(3).max(128).optional(),
-
-  // TODO: Type this properly from editorjs block types?
-  content: z.any().optional(),
-})
